requirements.txt
asgiref==3.2.10
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
deprecation==2.1.0
Django==3.0.8
django-sslserver==0.22
idna==2.10
jsonpickle==0.9.6
msgpack==1.0.0
packaging==20.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
squareup==6.0.0.20200625
urllib3==1.25.9
